added completly future to delete accaunt with apertunity of reactivate accaunt

+++ users/tasks.py
@@ -3,10 +3,11 @@ from .models import User
 
 @shared_task
 def hard_delete_user(user_id):
-    print("----in hard_delete_user")
+
     try:
         user_for_delete = User.objects.get(id=user_id)
-        user_for_delete.delete()
-        print("---------in hard_delete_user DELETED")
+        if not user_for_delete.is_active:
+            user_for_delete.delete()
+
     except User.DoesNotExist:
         pass
